Add more test coverage to Settings and Emitter.  Fix destroy bug when uncoupled

+++ src/main/utils/Settings.js
@@ -96,7 +96,7 @@ class Settings {
 
   destroy() {
     this.data = null;
-    fs.unlinkSync(this.PATH);
+    if (this.coupled && fs.existsSync(this.PATH)) fs.unlinkSync(this.PATH);
   }
 }
